dataset

'dataset' is an R6 class that stores nucleotide sequences, abundance, sample and treatment assignments, taxonomic classifications, asv / otu clusters and various reports. It is designed to facilitate data analysis across multiple R packages.

Author

Sarah Westcott, swestcot@umich.edu

Public fields

data

Rcpp::XPtr<Dataset> pointer to 'Dataset' c++ class. This allows package developers an easy access point to the underlying C++ code with additional functionality.

raw

Rcpp::RawVector containing the serialized data of the 'Dataset' c++ class. This allows the load and save functions to work with the class.

sequence_tree

a tree that relates sequences to eachother

sample_tree

a tree that relates samples to eachother

Methods

Public methods

dataset$new()

dataset$print()

dataset$add_sample_tree()

dataset$add_sequence_tree()

dataset$clear()

dataset$get_bin_assignments()

dataset$get_bin_names()

dataset$get_bin_types()

dataset$get_dataset_name()

dataset$get_metadata_table()

dataset$get_num_bins()

dataset$get_num_sequences()

dataset$get_num_treatments()

dataset$get_num_samples()

dataset$get_samples()

dataset$get_sample_tree()

dataset$get_scrap_report()

dataset$get_sequence_report()

dataset$get_summary()

dataset$get_sequence_tree()

dataset$get_treatments()

dataset$has_sample()

dataset$clone()

Method new()

Create a new dataset

Usage

dataset$new(
  name = "",
  processors = parallelly::availableCores(),
  dataset = NULL
)

Arguments

name

String, name of dataset (optional)

processors

Integer, number of cores to use. Default = all available

dataset

a `dataset` object.

Returns

A new `dataset` object.

Examples

# to create an empty dataset, run the following:

data <- new_dataset("soil")

Method print()

Get summary of sequence data

Usage

dataset$print()

Method add_sample_tree()

Add phylo tree relating the samples in your dataset

Usage

dataset$add_sample_tree(tree)

Arguments

tree

a phylo tree object created by ape::read.tree.

Examples

 data <- dataset$new("my_dataset")

 df <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))
 assign_bins(data, df)

 tree <- ape::read.tree(rdataset_example(
 "final.opti_mcc.jclass.ave.tre"))

 data$add_sample_tree(tree)

Method add_sequence_tree()

Add phylo tree relating the sequences in your dataset

Usage

dataset$add_sequence_tree(tree)

Arguments

tree

a phylo tree object created by ape::read.tree.

Examples

 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(rdataset_example("final.phylip.tre"))
 data$add_sequence_tree(tree)

Method clear()

Remove 'sample_tree', or 'sequence_tree' data from your dataset.

Usage

dataset$clear(tags = NULL)

Arguments

tags

a vector of strings containing the items you wish to clear. Options are 'metadata', 'references', 'sequence_tree', 'sample_tree', 'alignment_report', 'contigs_assembly_report' and '"'chimera_report'. By default, everything is cleared.

Method get_bin_assignments()

Get data frame containing sequence bin assignments

Usage

dataset$get_bin_assignments(type = "otu")

Arguments

type

a string indicating the type of clusters. Options include: "otu", "asv", or "phylotype". Default = "otu".

Returns

data.frame

Examples

  data <- dataset$new("my_dataset")

  bin_table <- readr::read_tsv(rdataset_example(
                               "mothur2_bin_assignments_shared.tsv"),
                               show_col_types = FALSE)

  assign_bins(data, bin_table)

  shared <- data$get_bin_assignments()

Method get_bin_names()

Get the names of the bins in your dataset

Usage

dataset$get_bin_names(type = "otu")

Arguments

type

a string indicating the type of clusters. Options include: "otu", "asv", or "phylotype". Default = "otu".

Returns

vector of strings containing the bin names

Examples

  data <- miseq_sop_example()
  data$get_bin_names("otu")

Method get_bin_types()

Get bin table types

Usage

dataset$get_bin_types()

Returns

vector of strings

Examples

data <- miseq_sop_example()
data$get_bin_types()

Method get_dataset_name()

Get dataset name

Usage

dataset$get_dataset_name()

Returns

String

Method get_metadata_table()

Get data.frame containing metadata for the dataset

Usage

dataset$get_metadata_table()

Returns

data.frame()

Examples

  data <- dataset$new("my_dataset")

  metadata <- readr::read_tsv(rdataset_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)

  add_metadata(data, metadata)

  data$get_metadata_table()

Method get_num_bins()

Get the number of bins in the dataset

Usage

dataset$get_num_bins(type = "otu")

Arguments

type

a string indicating the type of clusters. Default = "otu".

Returns

An integer

Examples

  otu_data <- read_mothur_list(rdataset_example(
                            "final.opti_mcc.list"))

  data <- dataset$new("my_dataset")
  assign_bins(data, otu_data)
  data$get_num_bins()

Method get_num_sequences()

Get the number of sequences in the dataset

Usage

dataset$get_num_sequences(distinct = FALSE, sample = NULL)

Arguments

distinct

Boolean. When distinct is TRUE the number of unique sequence is returned.

sample

The name of the sample you want number of sequences for, optional

Returns

An integer

Method get_num_treatments()

Get the number of treatments in the dataset

Usage

dataset$get_num_treatments()

Returns

An integer

Method get_num_samples()

Get the number of samples in the dataset

Usage

dataset$get_num_samples()

Returns

A character vector

Method get_samples()

Get names of samples in the dataset

Usage

dataset$get_samples()

Returns

A character vector

Method get_sample_tree()

Get phylo tree relating the samples in your dataset.

Usage

dataset$get_sample_tree()

Examples

 tree <- ape::read.tree(rdataset_example(
  "final.opti_mcc.jclass.ave.tre"))

 df <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))

 data <- dataset$new("my_dataset")
 assign_bins(data, df)
 data$add_sample_tree(tree)
 data$get_sample_tree()

Method get_scrap_report()

Get report containing the scrapped sequences / bins - ids, trash_codes

Usage

dataset$get_scrap_report()

Returns

list of data.frames

Method get_sequence_report()

Get data.frame sequence report data. Sequence report data includes: start positions, end positions, number of bases, number of ambiguous bases, length of longest homopolymer, and the number of N's.

Usage

dataset$get_sequence_report()

Returns

data.frame

Method get_summary()

Get summary of the sequence reports

Usage

dataset$get_summary(silent = FALSE)

Arguments

silent

Default = FALSE, meaning print summaries

Returns

list of data.frames

Method get_sequence_tree()

Get phylo tree relating the sequences in your dataset.

Usage

dataset$get_sequence_tree()

Examples

 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(rdataset_example("final.phylip.tre"))
 data$add_sequence_tree(tree)
 data$get_sequence_tree()

Method get_treatments()

Get names of treatments in the dataset

Usage

dataset$get_treatments()

Returns

A character vector

Method has_sample()

Determine if a sample is present in the dataset

Usage

dataset$has_sample(sample)

Arguments

sample

String, Name of sample

Returns

Boolean

Method clone()

The objects of this class are cloneable with this method.

Usage

dataset$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `dataset$new`
## ------------------------------------------------

# to create an empty dataset, run the following:

data <- new_dataset("soil")

## ------------------------------------------------
## Method `dataset$add_sample_tree`
## ------------------------------------------------

 data <- dataset$new("my_dataset")

 df <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))
 assign_bins(data, df)
#> ℹ Assigned 531 otu bins.
#> [1] 531

 tree <- ape::read.tree(rdataset_example(
 "final.opti_mcc.jclass.ave.tre"))

 data$add_sample_tree(tree)

## ------------------------------------------------
## Method `dataset$add_sequence_tree`
## ------------------------------------------------

 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(rdataset_example("final.phylip.tre"))
 data$add_sequence_tree(tree)
#> ℹ Added 2425 sequences.

## ------------------------------------------------
## Method `dataset$get_bin_assignments`
## ------------------------------------------------

  data <- dataset$new("my_dataset")

  bin_table <- readr::read_tsv(rdataset_example(
                               "mothur2_bin_assignments_shared.tsv"),
                               show_col_types = FALSE)

  assign_bins(data, bin_table)
#> ℹ Assigned 531 otu bins.
#> [1] 531

  shared <- data$get_bin_assignments()

## ------------------------------------------------
## Method `dataset$get_bin_names`
## ------------------------------------------------

  data <- miseq_sop_example()
#> ℹ Added 2425 sequences.
#> ℹ Assigned 2425 sequence abundances.
#> ℹ Assigned 2425 sequence taxonomies.
#> ℹ Assigned 531 otu bins.
#> ℹ Assigned 2425 asv bins.
#> ℹ Assigned 63 phylotype bins.
#> ℹ Assigned 19 samples to treatments.
#> ℹ Assigned 531 otu bin taxonomies.
#> ℹ Added 2 resource references.
#> ℹ Assigned 531 otu bin representative sequences.
  data$get_bin_names("otu")
#>   [1] "Otu001" "Otu002" "Otu003" "Otu004" "Otu005" "Otu006" "Otu007" "Otu008"
#>   [9] "Otu009" "Otu010" "Otu011" "Otu012" "Otu013" "Otu014" "Otu015" "Otu016"
#>  [17] "Otu017" "Otu018" "Otu019" "Otu020" "Otu021" "Otu022" "Otu023" "Otu024"
#>  [25] "Otu025" "Otu026" "Otu027" "Otu028" "Otu029" "Otu030" "Otu031" "Otu032"
#>  [33] "Otu033" "Otu034" "Otu035" "Otu036" "Otu037" "Otu038" "Otu039" "Otu040"
#>  [41] "Otu041" "Otu042" "Otu043" "Otu044" "Otu045" "Otu046" "Otu047" "Otu048"
#>  [49] "Otu049" "Otu050" "Otu051" "Otu052" "Otu053" "Otu054" "Otu055" "Otu056"
#>  [57] "Otu057" "Otu058" "Otu059" "Otu060" "Otu061" "Otu062" "Otu063" "Otu064"
#>  [65] "Otu065" "Otu066" "Otu067" "Otu068" "Otu069" "Otu070" "Otu071" "Otu072"
#>  [73] "Otu073" "Otu074" "Otu075" "Otu076" "Otu077" "Otu078" "Otu079" "Otu080"
#>  [81] "Otu081" "Otu082" "Otu083" "Otu084" "Otu085" "Otu086" "Otu087" "Otu088"
#>  [89] "Otu089" "Otu090" "Otu091" "Otu092" "Otu093" "Otu094" "Otu095" "Otu096"
#>  [97] "Otu097" "Otu098" "Otu099" "Otu100" "Otu101" "Otu102" "Otu103" "Otu104"
#> [105] "Otu105" "Otu106" "Otu107" "Otu108" "Otu109" "Otu110" "Otu111" "Otu112"
#> [113] "Otu113" "Otu114" "Otu115" "Otu116" "Otu117" "Otu118" "Otu119" "Otu120"
#> [121] "Otu121" "Otu122" "Otu123" "Otu124" "Otu125" "Otu126" "Otu127" "Otu128"
#> [129] "Otu129" "Otu130" "Otu131" "Otu132" "Otu133" "Otu134" "Otu135" "Otu136"
#> [137] "Otu137" "Otu138" "Otu139" "Otu140" "Otu141" "Otu142" "Otu143" "Otu144"
#> [145] "Otu145" "Otu146" "Otu147" "Otu148" "Otu149" "Otu150" "Otu151" "Otu152"
#> [153] "Otu153" "Otu154" "Otu155" "Otu156" "Otu157" "Otu158" "Otu159" "Otu160"
#> [161] "Otu161" "Otu162" "Otu163" "Otu164" "Otu165" "Otu166" "Otu167" "Otu168"
#> [169] "Otu169" "Otu170" "Otu171" "Otu172" "Otu173" "Otu174" "Otu175" "Otu176"
#> [177] "Otu177" "Otu178" "Otu179" "Otu180" "Otu181" "Otu182" "Otu183" "Otu184"
#> [185] "Otu185" "Otu186" "Otu187" "Otu188" "Otu189" "Otu190" "Otu191" "Otu192"
#> [193] "Otu193" "Otu194" "Otu195" "Otu196" "Otu197" "Otu198" "Otu199" "Otu200"
#> [201] "Otu201" "Otu202" "Otu203" "Otu204" "Otu205" "Otu206" "Otu207" "Otu208"
#> [209] "Otu209" "Otu210" "Otu211" "Otu212" "Otu213" "Otu214" "Otu215" "Otu216"
#> [217] "Otu217" "Otu218" "Otu219" "Otu220" "Otu221" "Otu222" "Otu223" "Otu224"
#> [225] "Otu225" "Otu226" "Otu227" "Otu228" "Otu229" "Otu230" "Otu231" "Otu232"
#> [233] "Otu233" "Otu234" "Otu235" "Otu236" "Otu237" "Otu238" "Otu239" "Otu240"
#> [241] "Otu241" "Otu242" "Otu243" "Otu244" "Otu245" "Otu246" "Otu247" "Otu248"
#> [249] "Otu249" "Otu250" "Otu251" "Otu252" "Otu253" "Otu254" "Otu255" "Otu256"
#> [257] "Otu257" "Otu258" "Otu259" "Otu260" "Otu261" "Otu262" "Otu263" "Otu264"
#> [265] "Otu265" "Otu266" "Otu267" "Otu268" "Otu269" "Otu270" "Otu271" "Otu272"
#> [273] "Otu273" "Otu274" "Otu275" "Otu276" "Otu277" "Otu278" "Otu279" "Otu280"
#> [281] "Otu281" "Otu282" "Otu283" "Otu284" "Otu285" "Otu286" "Otu287" "Otu288"
#> [289] "Otu289" "Otu290" "Otu291" "Otu292" "Otu293" "Otu294" "Otu295" "Otu296"
#> [297] "Otu297" "Otu298" "Otu299" "Otu300" "Otu301" "Otu302" "Otu303" "Otu304"
#> [305] "Otu305" "Otu306" "Otu307" "Otu308" "Otu309" "Otu310" "Otu311" "Otu312"
#> [313] "Otu313" "Otu314" "Otu315" "Otu316" "Otu317" "Otu318" "Otu319" "Otu320"
#> [321] "Otu321" "Otu322" "Otu323" "Otu324" "Otu325" "Otu326" "Otu327" "Otu328"
#> [329] "Otu329" "Otu330" "Otu331" "Otu332" "Otu333" "Otu334" "Otu335" "Otu336"
#> [337] "Otu337" "Otu338" "Otu339" "Otu340" "Otu341" "Otu342" "Otu343" "Otu344"
#> [345] "Otu345" "Otu346" "Otu347" "Otu348" "Otu349" "Otu350" "Otu351" "Otu352"
#> [353] "Otu353" "Otu354" "Otu355" "Otu356" "Otu357" "Otu358" "Otu359" "Otu360"
#> [361] "Otu361" "Otu362" "Otu363" "Otu364" "Otu365" "Otu366" "Otu367" "Otu368"
#> [369] "Otu369" "Otu370" "Otu371" "Otu372" "Otu373" "Otu374" "Otu375" "Otu376"
#> [377] "Otu377" "Otu378" "Otu379" "Otu380" "Otu381" "Otu382" "Otu383" "Otu384"
#> [385] "Otu385" "Otu386" "Otu387" "Otu388" "Otu389" "Otu390" "Otu391" "Otu392"
#> [393] "Otu393" "Otu394" "Otu395" "Otu396" "Otu397" "Otu398" "Otu399" "Otu400"
#> [401] "Otu401" "Otu402" "Otu403" "Otu404" "Otu405" "Otu406" "Otu407" "Otu408"
#> [409] "Otu409" "Otu410" "Otu411" "Otu412" "Otu413" "Otu414" "Otu415" "Otu416"
#> [417] "Otu417" "Otu418" "Otu419" "Otu420" "Otu421" "Otu422" "Otu423" "Otu424"
#> [425] "Otu425" "Otu426" "Otu427" "Otu428" "Otu429" "Otu430" "Otu431" "Otu432"
#> [433] "Otu433" "Otu434" "Otu435" "Otu436" "Otu437" "Otu438" "Otu439" "Otu440"
#> [441] "Otu441" "Otu442" "Otu443" "Otu444" "Otu445" "Otu446" "Otu447" "Otu448"
#> [449] "Otu449" "Otu450" "Otu451" "Otu452" "Otu453" "Otu454" "Otu455" "Otu456"
#> [457] "Otu457" "Otu458" "Otu459" "Otu460" "Otu461" "Otu462" "Otu463" "Otu464"
#> [465] "Otu465" "Otu466" "Otu467" "Otu468" "Otu469" "Otu470" "Otu471" "Otu472"
#> [473] "Otu473" "Otu474" "Otu475" "Otu476" "Otu477" "Otu478" "Otu479" "Otu480"
#> [481] "Otu481" "Otu482" "Otu483" "Otu484" "Otu485" "Otu486" "Otu487" "Otu488"
#> [489] "Otu489" "Otu490" "Otu491" "Otu492" "Otu493" "Otu494" "Otu495" "Otu496"
#> [497] "Otu497" "Otu498" "Otu499" "Otu500" "Otu501" "Otu502" "Otu503" "Otu504"
#> [505] "Otu505" "Otu506" "Otu507" "Otu508" "Otu509" "Otu510" "Otu511" "Otu512"
#> [513] "Otu513" "Otu514" "Otu515" "Otu516" "Otu517" "Otu518" "Otu519" "Otu520"
#> [521] "Otu521" "Otu522" "Otu523" "Otu524" "Otu525" "Otu526" "Otu527" "Otu528"
#> [529] "Otu529" "Otu530" "Otu531"

## ------------------------------------------------
## Method `dataset$get_bin_types`
## ------------------------------------------------

data <- miseq_sop_example()
#> ℹ Added 2425 sequences.
#> ℹ Assigned 2425 sequence abundances.
#> ℹ Assigned 2425 sequence taxonomies.
#> ℹ Assigned 531 otu bins.
#> ℹ Assigned 2425 asv bins.
#> ℹ Assigned 63 phylotype bins.
#> ℹ Assigned 19 samples to treatments.
#> ℹ Assigned 531 otu bin taxonomies.
#> ℹ Added 2 resource references.
#> ℹ Assigned 531 otu bin representative sequences.
data$get_bin_types()
#> [1] "otu"       "asv"       "phylotype"

## ------------------------------------------------
## Method `dataset$get_metadata_table`
## ------------------------------------------------

  data <- dataset$new("my_dataset")

  metadata <- readr::read_tsv(rdataset_example("sample-metadata.tsv"),
   col_names = TRUE, show_col_types = FALSE)

  add_metadata(data, metadata)

  data$get_metadata_table()
#>    sample-id barcode-sequence  body-site year month day   subject
#> 1       L1S8     AGCTGACTAGTC        gut 2008    10  28 subject-1
#> 2      L1S57     ACACACTATGGC        gut 2009     1  20 subject-1
#> 3      L1S76     ACTACGTGTGGT        gut 2009     2  17 subject-1
#> 4     L1S105     AGTGCGATGCGT        gut 2009     3  17 subject-1
#> 5     L2S155     ACGATGCGACCA  left palm 2009     1  20 subject-1
#> 6     L2S175     AGCTATCCACGA  left palm 2009     2  17 subject-1
#> 7     L2S204     ATGCAGCTCAGT  left palm 2009     3  17 subject-1
#> 8     L2S222     CACGTGACATGT  left palm 2009     4  14 subject-1
#> 9     L3S242     ACAGTTGCGCGA right palm 2008    10  28 subject-1
#> 10    L3S294     CACGACAGGCTA right palm 2009     1  20 subject-1
#> 11    L3S313     AGTGTCACGGTG right palm 2009     2  17 subject-1
#> 12    L3S341     CAAGTGAGAGAG right palm 2009     3  17 subject-1
#> 13    L3S360     CATCGTATCAAC right palm 2009     4  14 subject-1
#> 14    L5S104     CAGTGTCAGGAC     tongue 2008    10  28 subject-1
#> 15    L5S155     ATCTTAGACTGC     tongue 2009     1  20 subject-1
#> 16    L5S174     CAGACATTGCGT     tongue 2009     2  17 subject-1
#> 17    L5S203     CGATGCACCAGA     tongue 2009     3  17 subject-1
#> 18    L5S222     CTAGAGACTCTT     tongue 2009     4  14 subject-1
#> 19    L1S140     ATGGCAGCTCTA        gut 2008    10  28 subject-2
#> 20    L1S208     CTGAGATACGCG        gut 2009     1  20 subject-2
#> 21    L1S257     CCGACTGAGATG        gut 2009     3  17 subject-2
#> 22    L1S281     CCTCTCGTGATC        gut 2009     4  14 subject-2
#> 23    L2S240     CATATCGCAGTT  left palm 2008    10  28 subject-2
#> 24    L2S309     CGTGCATTATCA  left palm 2009     1  20 subject-2
#> 25    L2S357     CTAACGCAGTCA  left palm 2009     3  17 subject-2
#> 26    L2S382     CTCAATGACTCA  left palm 2009     4  14 subject-2
#> 27    L3S378     ATCGATCTGTGG right palm 2008    10  28 subject-2
#> 28     L4S63     CTCGTGGAGTAG right palm 2009     1  20 subject-2
#> 29    L4S112     GCGTTACACACA right palm 2009     3  17 subject-2
#> 30    L4S137     GAACTGTATCTC right palm 2009     4  14 subject-2
#> 31    L5S240     CTGGACTCATAG     tongue 2008    10  28 subject-2
#> 32     L6S20     GAGGCTCATCAT     tongue 2009     1  20 subject-2
#> 33     L6S68     GATACGTCCTGA     tongue 2009     3  17 subject-2
#> 34     L6S93     GATTAGCACTCT     tongue 2009     4  14 subject-2
#>    reported-antibiotic-usage days-since-experiment-start
#> 1                        Yes                           0
#> 2                         No                          84
#> 3                         No                         112
#> 4                         No                         140
#> 5                         No                          84
#> 6                         No                         112
#> 7                         No                         140
#> 8                         No                         168
#> 9                        Yes                           0
#> 10                        No                          84
#> 11                        No                         112
#> 12                        No                         140
#> 13                        No                         168
#> 14                       Yes                           0
#> 15                        No                          84
#> 16                        No                         112
#> 17                        No                         140
#> 18                        No                         168
#> 19                       Yes                           0
#> 20                        No                          84
#> 21                        No                         140
#> 22                        No                         168
#> 23                       Yes                           0
#> 24                        No                          84
#> 25                        No                         140
#> 26                        No                         168
#> 27                       Yes                           0
#> 28                        No                          84
#> 29                        No                         140
#> 30                        No                         168
#> 31                       Yes                           0
#> 32                        No                          84
#> 33                        No                         140
#> 34                        No                         168

## ------------------------------------------------
## Method `dataset$get_num_bins`
## ------------------------------------------------

  otu_data <- read_mothur_list(rdataset_example(
                            "final.opti_mcc.list"))

  data <- dataset$new("my_dataset")
  assign_bins(data, otu_data)
#> ℹ Assigned 531 otu bins.
#> [1] 531
  data$get_num_bins()
#> [1] 531

## ------------------------------------------------
## Method `dataset$get_sample_tree`
## ------------------------------------------------

 tree <- ape::read.tree(rdataset_example(
  "final.opti_mcc.jclass.ave.tre"))

 df <- read_mothur_shared(rdataset_example("final.opti_mcc.shared"))

 data <- dataset$new("my_dataset")
 assign_bins(data, df)
#> ℹ Assigned 531 otu bins.
#> [1] 531
 data$add_sample_tree(tree)
 data$get_sample_tree()
#> 
#> Phylogenetic tree with 19 tips and 18 internal nodes.
#> 
#> Tip labels:
#>   F3D9, F3D8, F3D6, F3D5, F3D2, F3D1, ...
#> 
#> Rooted; includes branch length(s).

## ------------------------------------------------
## Method `dataset$get_sequence_tree`
## ------------------------------------------------

 data <- dataset$new("my_dataset")
 tree <- ape::read.tree(rdataset_example("final.phylip.tre"))
 data$add_sequence_tree(tree)
#> ℹ Added 2425 sequences.
 data$get_sequence_tree()
#> 
#> Phylogenetic tree with 2425 tips and 2424 internal nodes.
#> 
#> Tip labels:
#>   M00967_43_000000000-A3JHG_1_1114_15727_25995, M00967_43_000000000-A3JHG_1_2109_19976_22044, M00967_43_000000000-A3JHG_1_1102_9244_9305, M00967_43_000000000-A3JHG_1_2101_14159_9619, M00967_43_000000000-A3JHG_1_1111_12315_7486, M00967_43_000000000-A3JHG_1_1107_12586_26826, ...
#> 
#> Rooted; includes branch length(s).