get_bin

Get the names of the sequences in a given bin in a dataset object

get_bin(data, bin_name, type = "otu")

Arguments

data,

a dataset object

bin_name,

string containing the bin name

type

a string indicating the type of clusters. Default = "otu".

Value

String, containing names of the sequences in a given bin

Examples

data <- miseq_sop_example()
#> ℹ Added 2425 sequences.
#> ℹ Assigned 2425 sequence abundances.
#> ℹ Assigned 2425 sequence taxonomies.
#> ℹ Assigned 531 otu bins.
#> ℹ Assigned 2425 asv bins.
#> ℹ Assigned 63 phylotype bins.
#> ℹ Assigned 19 samples to treatments.
#> ℹ Assigned 531 otu bin taxonomies.
#> ℹ Added 2 resource references.
#> ℹ Assigned 531 otu bin representative sequences.
get_bin(data, "Otu005", "otu")
#> [1] "M00967_43_000000000-A3JHG_1_2107_5816_12823,M00967_43_000000000-A3JHG_1_1112_17779_24453,M00967_43_000000000-A3JHG_1_1111_17856_12278,M00967_43_000000000-A3JHG_1_2111_29227_17392,M00967_43_000000000-A3JHG_1_1108_11375_17687,M00967_43_000000000-A3JHG_1_2109_20497_25999,M00967_43_000000000-A3JHG_1_2102_8178_6298,M00967_43_000000000-A3JHG_1_1113_25843_7710,M00967_43_000000000-A3JHG_1_2101_25853_16524,M00967_43_000000000-A3JHG_1_1110_12935_6977,M00967_43_000000000-A3JHG_1_2107_25612_9862,M00967_43_000000000-A3JHG_1_1102_15600_19124,M00967_43_000000000-A3JHG_1_1102_26826_15050,M00967_43_000000000-A3JHG_1_1101_23090_14223,M00967_43_000000000-A3JHG_1_2104_4792_21056,M00967_43_000000000-A3JHG_1_1113_15543_24961,M00967_43_000000000-A3JHG_1_2111_6782_8678,M00967_43_000000000-A3JHG_1_1106_16049_12791,M00967_43_000000000-A3JHG_1_1106_5620_18651,M00967_43_000000000-A3JHG_1_2113_10242_4487,M00967_43_000000000-A3JHG_1_2108_18302_14612,M00967_43_000000000-A3JHG_1_1108_19974_13521,M00967_43_000000000-A3JHG_1_1110_8577_13143,M00967_43_000000000-A3JHG_1_2108_18545_16099,M00967_43_000000000-A3JHG_1_1105_13025_22602,M00967_43_000000000-A3JHG_1_2106_19750_8249,M00967_43_000000000-A3JHG_1_1114_26027_6706,M00967_43_000000000-A3JHG_1_1104_13037_4245,M00967_43_000000000-A3JHG_1_2108_17961_17141,M00967_43_000000000-A3JHG_1_2110_22767_24294,M00967_43_000000000-A3JHG_1_1113_13105_21336,M00967_43_000000000-A3JHG_1_2104_11018_16527,M00967_43_000000000-A3JHG_1_1112_14201_3200,M00967_43_000000000-A3JHG_1_2107_23464_17422,M00967_43_000000000-A3JHG_1_2110_16698_16415,M00967_43_000000000-A3JHG_1_1113_25861_7699,M00967_43_000000000-A3JHG_1_1106_25389_22814,M00967_43_000000000-A3JHG_1_1111_15521_13523,M00967_43_000000000-A3JHG_1_2113_5393_13669,M00967_43_000000000-A3JHG_1_1109_7971_24259,M00967_43_000000000-A3JHG_1_1110_8805_12834,M00967_43_000000000-A3JHG_1_1104_19589_20033,M00967_43_000000000-A3JHG_1_2101_26706_7767,M00967_43_000000000-A3JHG_1_2114_4609_11919,M00967_43_000000000-A3JHG_1_1110_6233_13348,M00967_43_000000000-A3JHG_1_1104_23236_5434,M00967_43_000000000-A3JHG_1_2104_14582_18801,M00967_43_000000000-A3JHG_1_2104_26095_18712,M00967_43_000000000-A3JHG_1_2108_23551_13901,M00967_43_000000000-A3JHG_1_1103_19447_15932,M00967_43_000000000-A3JHG_1_1102_15632_13472,M00967_43_000000000-A3JHG_1_1105_17602_8360,M00967_43_000000000-A3JHG_1_1104_10411_22472,M00967_43_000000000-A3JHG_1_2107_22192_18692,M00967_43_000000000-A3JHG_1_1112_16673_28539,M00967_43_000000000-A3JHG_1_1107_15750_18592"