get_sequence_abundances

Get the total abundance for each sequence in a dataset object

get_sequence_abundances(data)

Arguments

data,

a dataset object

Value

vector containing the total abundance for each sequence a dataset object

Examples

data <- miseq_sop_example()
#> ℹ Added 2425 sequences.
#> ℹ Assigned 2425 sequence abundances.
#> ℹ Assigned 2425 sequence taxonomies.
#> ℹ Assigned 531 otu bins.
#> ℹ Assigned 2425 asv bins.
#> ℹ Assigned 63 phylotype bins.
#> ℹ Assigned 19 samples to treatments.
#> ℹ Assigned 531 otu bin taxonomies.
#> ℹ Added 2 resource references.
#> ℹ Assigned 531 otu bin representative sequences.
get_sequence_abundances(data)
#>    [1]     1     1     1     1     1     1   191     1     1     1     1     5
#>   [13]     1     2     1     1     1     7     1     1     1     1     1     1
#>   [25]     1     1     4     1     1     1     2     4     1     1     1     1
#>   [37]     1     1     1     8     1     1     1     1     4    20     1     1
#>   [49]     1     9     3     1     1    90   156     1     1     1     1     1
#>   [61]     1     1     7     1     1     1     1     1     1     1     1    37
#>   [73]     1     8     1   620     1     1   136     1     5     2     1     1
#>   [85]    71     1     5     1     1     1     1     1     1     1     1     2
#>   [97]     1     1     1     1     6     1     1     1     1     1     1     1
#>  [109]     8     1     1     1     1     1     1     1     1     1     1     1
#>  [121]     6     1     1     2     1     1     1     1     1    31     1     1
#>  [133]     1     3     1     4     1     1     1     1     1     1     1     1
#>  [145]   178     1     1     1     1     1     1     1     1     1   131     1
#>  [157]    14     1     1     1     1     3     1     1     2     1     1     1
#>  [169]     1     1     1     1     2     1     1     1     1     1  6285     1
#>  [181]     1     1     1     1     1     1     1    44     1     1     2     9
#>  [193]     1     1     1   248   315     1     7     1     1     1     1     1
#>  [205]     1     1     1     1     1     1     1     1     2    28     1     1
#>  [217]     1     1     1     3     1    16     1   251     3     1     2     1
#>  [229]     2     1     1    10     2     1     1     1   195     1     1     1
#>  [241]     1     1     1     1     4     1    33     1     1     1     1     1
#>  [253]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [265]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [277]     1     1     1     1    11     1     1     1     1     1     6     1
#>  [289]     1     1     1     1     1     1   334    16     1     1     1     1
#>  [301]   235     1     1     1     1     1     1     9     1     1     1     1
#>  [313]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [325]     1     1     8     1     1    17    89     1     1     1     2     1
#>  [337]     2     1     1     1     1     1     1    26     1     1     1     1
#>  [349]     1     1     1     1     1     1     1     1     1    75     3     1
#>  [361]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [373]     1     7     2     7     8     1     1     1     1     1     1     1
#>  [385]     1     1     2     1     1     1     2     1     1     1   386    93
#>  [397]     2     1     1     1     1     1     1     1     1     1     2     1
#>  [409]     4     1     1     1     1     1     1     1     1     1   131    77
#>  [421]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [433]     8     1     1     1     1     1     1     1     1     1     1     1
#>  [445]     1     1     6     1     1     1     1     1     1     1     1     1
#>  [457]     1     2     1     1     1     1     1     1     1   222     2     1
#>  [469]     1     1     1     1     1     1     1     1     1     1     1    17
#>  [481]     1     1     1     1     1     1     1     1     2     1     1     1
#>  [493]     1     1     1     1     1     1     1     1     1     1    17    22
#>  [505]     1    71     1     1     1     1     2    13     1     1     1     1
#>  [517]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [529]     1     1    80    14     1     1     2     1    37     1   820     1
#>  [541]     1     1     1     1     1     1     3    22     1     1    35     1
#>  [553]   708     1     1     1     1     1     1     1    33     1     1     1
#>  [565]     1     1     1     1     1     1     1     2     2     1     1     1
#>  [577]     1    30     1   490     1     1     1     1     1     1     1     1
#>  [589]     1     1     1    21     1     1     1     1     1     1     2     1
#>  [601]     1     1     1     1     1     1     1     7    31     1     1     1
#>  [613]     6     1     1     1     1     1     1     2    41     1     1     1
#>  [625]     1     1     1     1  3565     1     1     1     1     1     1     1
#>  [637]     1     1     1     1    12     1     1     1     1     1     1     1
#>  [649]     6     2     1     1     1     2     4     1     1     1     3     1
#>  [661]     1     1     1     2     1     1     1     1     1     1     2     1
#>  [673]     1     1     1     1     1     1     2     1     4     1     1     1
#>  [685]     1     9     9   375     3     1     1     1     1     1     1     1
#>  [697]     2     1     1     1     1     1     1     1     1     1     1     2
#>  [709]     1    47     1     1     3     1     1     1     1     8     1     1
#>  [721]     2     1     1    39    51   791     1     1     1    23     1     1
#>  [733]     3    17     4     3     1     1     1     1     3     1     1     1
#>  [745]     1     1    10     1     1     1     1     1     2     5     1     1
#>  [757]   221     1     1     1     1     2     1     1     4     1     1     1
#>  [769]     1     1     1     1     2     5    11   144   130     1     1     1
#>  [781]     1     8     1     1     1     1     1     1     1     1     1     1
#>  [793]     1     2     1     1     1     1    21     1     1     2    33     9
#>  [805]     1     1   277     1     1     1     1     1     1     1   145     1
#>  [817]     1     1     1     1     1     1    20     9     1     1     1     1
#>  [829]     1     1     1     1     1     1     1     1     1     1    32     1
#>  [841]     1     1     1     1     1     1     1     1     1    69     1     1
#>  [853]     1   505     1     1     8     1     1     1     1     1     1     1
#>  [865]     1     1    92     1     1     1     1     1     1   697     1     1
#>  [877]     1     1     1     3     1     1     1     1     1     1     1     1
#>  [889]     1     1     1     2     1     6     1     1    18     1     1     1
#>  [901]     1     1     1     1     1     1     1     1     1     1   365     1
#>  [913]     1     1   191     1     1     1     3   961     1     1     1     1
#>  [925]     1    12     1     1     2     1     1     1     2     1     1     1
#>  [937]     1     1     1     2     1     1     1     1     1     1     1     2
#>  [949]     2     1     1    17     4     8     1     1     1     1  1272     1
#>  [961]     1     1     2     1     1     1     1     1     1     1     1    74
#>  [973]     1    13     1     5     7     1     1     1     1     1    98     1
#>  [985]     1     1     1   369     1     1     1     1     1     1     5     1
#>  [997]     1     1     1   188     1    16     1     1     1     2     1     5
#> [1009]     1   157     1     1     1     1     1     1     2     5     1     1
#> [1021]    11     1   117     1     1     1   303     1     1     1     1     1
#> [1033]     1     1     1     1     1     1  7698     2     1     2   107     1
#> [1045]     1     5     1     1     1     1     1     1     1     1     1     1
#> [1057]     1     1     1     1     4     5     1     6     1     1     2    13
#> [1069]     1     1     2     2     6     1     1     1     1     1     1     1
#> [1081]     1     1   100    10     1  5207     9     1     1   225     1     1
#> [1093]     1     1     1     1     1     1     3     1     1    20    60     1
#> [1105]    35   286     1     1     1     1   220     1     1     1     1     1
#> [1117]     1     1     1     1     2     3     1   243     1  7436     1     1
#> [1129]     1     1     1     1     1     1     1     1     1     5     1     1
#> [1141]     1     1     1     8     1     1     1   380     1     1     1     1
#> [1153]     1    21     9     1     1     1     1     1     1     1     1     3
#> [1165]     2     1     1     1     1     1    16     4     1     1     8     1
#> [1177]     4     1     1     1     1     1     1     1     1     1    40     1
#> [1189]     1     1     1     1     1     6     5     1     1     5     1     1
#> [1201]     1     1     1     1   199     1     1     1     1     2     1     1
#> [1213]     2     1     1     6     1     1     1     1     1     1   209     5
#> [1225]     1     1     1     1     1     1     1     1     1  1089     1     1
#> [1237]   930     1     1     1     1     1     1     1     6     1     1     1
#> [1249]     1     1     1     1     1    99     1     1     1     1     1    43
#> [1261]     1     1     1     2     1     2     1     1     1     1     1    47
#> [1273]     1     4     1     1     1     1     1     1     1     1     1     1
#> [1285]   297     1     1     1     1     1     1     1     1     1     1     1
#> [1297]     1     1     1     1     1     1     1     1   560     2     1     1
#> [1309]     1     7     1     1     1     1     1     1    57     1     1    46
#> [1321]     1     1     1     1     1     2     1     1     1     1     1     1
#> [1333]     1     1     1     1     4     1   586     1     1     1     1     1
#> [1345]     1     1     1     1     1     1     1     1     1     1     2     1
#> [1357]     1     2     1     1     1     1     1     1     1     1     1     1
#> [1369]     1    35     1     1     1     1     1     1     1  2854   222     1
#> [1381]     1     2     1     1     2    39     1     2     1     1     1     1
#> [1393]     1    21    15     1     3     1     1     1     1     2     1     1
#> [1405]     1    16     1     1     1     1     1     1     1     1     1     1
#> [1417]     1     1     1     1     1     1     1     1     1     1     3    98
#> [1429]     1     1     1     1     1     1     1     1     1     1     1     1
#> [1441]     1     1     1     1     1     1     1     2     1     1   132     1
#> [1453]     1  1129     1     1     1     3     1     1     1     1     1     1
#> [1465]     3     1     1     6     3     1     1     1    43   103     1     1
#> [1477]     1     2     1     1   451     1     1     1     1     1     1    70
#> [1489]     1     1     5     1     1     1     2     1     1    90     1     1
#> [1501]     1     1     3     1     1     1     1    10     1     1     1    11
#> [1513]     1     1     1     1     1     1     1     1    57     1     1     9
#> [1525]     1     2     1     1  3050     1     1     2     1     8     3    76
#> [1537]     1     1     1    48     1     1     1     1     1     1     1     1
#> [1549]     1    10   136     1     1     1     1     1     1     1     1     1
#> [1561]     1     1     1     1     3     1     1     1     1     1     9     1
#> [1573]     1     1     1     1     1     1     1     1    21     1     5     6
#> [1585]     1     1     1     1   205     1     1     1     1     1     3     1
#> [1597]     1     1     1     1     1     1     1     1    24     1    26    85
#> [1609]     1     1     1     1     1     1  1064     1     1     5     1     2
#> [1621]     1     3     2     1     1     2   206     1     1     1     1   331
#> [1633]     1     7    33    11    74     1     1     1     1     3     1     1
#> [1645]     1     1     1     1     1     1     1     3    47     1     1     1
#> [1657]     1     7     2     2     1     1     1     1     1   282    15     1
#> [1669]     1     1     1     1     1     1   633     1    57     1     1     1
#> [1681]     1     2     1     1   108     1     1     1    46     1     2     1
#> [1693]     1     1     1     2     1     1     1     2     1     1     2     5
#> [1705]   169     1     1    18     1     1     1     1     8     1     1     1
#> [1717]     1     2     1     1     1     5     1    25     1     4    35     1
#> [1729]     1     1     2     1     1     1     1     1     1  1958     1     1
#> [1741]   210   520     1     1     1    12     1     1     1     5     1    43
#> [1753]     1     1     1     1     1    30   182     1     1     2     1     1
#> [1765]     1     1     1     1     1     1     5     1    27     1     1     1
#> [1777]     1     4     1     1     1    69     1     1     1     1     1     1
#> [1789]     1     5     1     1     1    62     3     1     1     1   656     1
#> [1801]     1     1     1     1     1     8     1     1     1     1     1     1
#> [1813]     1     2     1     1     1     1     1     1     1     1     1    24
#> [1825]     1     2     1     1   210     5     2     1     1     1     1     1
#> [1837]     1     1    51     1     1     1    32     1     1     1     1     1
#> [1849]     1     1     1     1     1     2     1     1     1     1    12   350
#> [1861]     1     1     2     1    21     1     1     1     1    77     1     1
#> [1873]     1     1    14     1     1     1    21    10    11     1     1     1
#> [1885]     3     1     1     9     1     1     1     1     1     1     2     1
#> [1897]     1     1     1     1     2     2     1     3     1     1     1     5
#> [1909]     1     1     2     1     3     1     1     7     1     1     1     1
#> [1921]     1     1     1     1     1     1     1     1     6     1     1     1
#> [1933]     1     1     1     1     1     1     1   335     1     1     1     1
#> [1945]     1     2     1     1     1     1     2     3     2     4     1     1
#> [1957]     1     1     1     1     1     1     4     2     2    15     1     1
#> [1969]   326     1     1     1     1     1     1     1     1     1     1     1
#> [1981]     1     3     1   615   137     1     3     1    24    73     1     3
#> [1993]     1     1     2     1     1     1     2     1     1     1     1     1
#> [2005]     8     1     2     1     1     1     1     1     1     1     4    17
#> [2017]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2029]     1     1    47 12196     1     1     1     1     1     1    55    13
#> [2041]     3     2     1     1     1     1     1     1  8829     1     1     2
#> [2053]     3     1     1     2     1     1   127     5     1     1     1     1
#> [2065]     1     1     1     1     1     1     1     1     4     1     9     1
#> [2077]     1     1     1     1    47     1     1     1     1     1     1  1297
#> [2089]     1     5     1     1    12    53     1     1     2    80     2     1
#> [2101]     1     1     1     1     2     1     1     1     2     1     1     1
#> [2113]     1     1     1     1    62     1     1    11     1     1     1     1
#> [2125]     1     1     1     2     1     1     2     1     1     1     1     1
#> [2137]     1     1     1     1     1     1     1    61     1     1     1     1
#> [2149]     1     1     1     1     1     1     1     1     1     1     1     8
#> [2161]     1     1     1     1     1    47    64     1     1     1     1     1
#> [2173]     1     1     4     1     1     1    43     1     1     1     1    87
#> [2185]     1     1     1     1     1     1     1     1    11     1     4     1
#> [2197]     1     6     1     1     1     2    41     5     1     1     1     1
#> [2209]     1     1   359     1     1     1     2     2     1     1    37     1
#> [2221]     1     1     2     1     1     1     1     1   229     1     1    75
#> [2233]     1     1     1     1     1     1     1     1     8     3     1     1
#> [2245]     1     1     1     1  6484     1     1     1     1     1     2     1
#> [2257]     1     1     1  1835     1     1     1     1     4     1     1     3
#> [2269]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2281]     2     1     1     1     1     1     1     1     1     1     1     1
#> [2293]     1     1     1     7     1     1     6     1     1     1     1     1
#> [2305]     2     1     1     4     1     1     1     1    33    12     1     1
#> [2317]     1     1     1     1     1     1    19     1     1     1     1     2
#> [2329]     1     1     3     1     1     5     1     1     1     1     1     1
#> [2341]    24     1     1     1     1     1     1     1     1     1     1     1
#> [2353]     1     1     2     1     1     1     1     1     1     1     1    27
#> [2365]     1     1     3     1   320    16     9     1     2     1   425     1
#> [2377]     1     6     1     1     1     1     1     1     1     1     1     1
#> [2389]     1     1     1  7307     1   290     1     1     1     1     1     1
#> [2401]     1     1     1     1     1     1     1     1     1     1     1     1
#> [2413]     1     1     1     1     7     1     1   781     1     1     1    32
#> [2425]     1