get_shared_vector

Get 2D vector containing bin abundance data by sample in a dataset object

get_shared_vector(data, type = "otu")

Arguments

data,

a dataset object.

type

a string indicating the type of bin assignments. Default "otu".

Value

2D vector ([num_bins][num_samples]) containing the abundances of each bin parsed by sample.

Examples

 data <- miseq_sop_example()
#> ℹ Added 2425 sequences.
#> ℹ Assigned 2425 sequence abundances.
#> ℹ Assigned 2425 sequence taxonomies.
#> ℹ Assigned 531 otu bins.
#> ℹ Assigned 2425 asv bins.
#> ℹ Assigned 63 phylotype bins.
#> ℹ Assigned 19 samples to treatments.
#> ℹ Assigned 531 otu bin taxonomies.
#> ℹ Added 2 resource references.
#> ℹ Assigned 531 otu bin representative sequences.
 get_shared_vector(data)
#> [[1]]
#>  [1]  499  351  388  244  189  346  566  270 1316  750  770  265 3076  831  278
#> [16]  902  576  236  435
#> 
#> [[2]]
#>  [1]  306  311  335  258  152  243  445  204 1107  666  701  207 1439  543  235
#> [16]  608  449  309  374
#> 
#> [[3]]
#>  [1]  394  189  301  142  173  260  461  225  819  520  650  343 1065  400  236
#> [16]  507  385  295  429
#> 
#> [[4]]
#>  [1]  403   64  482  152  207  316  527  368 1052  830  861  436  451  187  148
#> [16]  381  294  128  189
#> 
#> [[5]]
#>  [1]  632   73  425  253  308  469  665  397 1619 1230  879  281  148   23   20
#> [16]   12    9    2    5
#> 
#> [[6]]
#>  [1]  356  117  279  193  183  282  471  257 1051  725  651  276  410  213  137
#> [16]  356  235  174  255
#> 
#> [[7]]
#>  [1] 168 127 206 201 116 132 362 193 545 499 454 182 354 448 142 501 517 547 610
#> 
#> [[8]]
#>  [1]  164  174  331   81   92   36  127   66   89  542  542  110 1222  398  193
#> [16]  250  223  278  419
#> 
#> [[9]]
#>  [1] 136  83 119  72  62 140 217  78 240 186 257 132 572 268 191 218 251 181 203
#> 
#> [[10]]
#>  [1]  19 111 175  43  76 256 319 185 498 419 497  57  60 175  67  48  13  15  28
#> 
#> [[11]]
#>  [1]  27  53 111  51  70  96 131  80 526 283 290 161 306 221 115 128 117  74  97
#> 
#> [[12]]
#>  [1] 110  35 101  81  43  84 188  71 476 239 159  71 143  65  66  59  43  18  22
#> 
#> [[13]]
#>  [1]  52 115   8  96  37  13  18   4 135  17  74  55 308  84  45 392 106 129 166
#> 
#> [[14]]
#>  [1] 104 252  13   6   2   7   8  39  22  47  39  19 550  16  83  72  25  50  98
#> 
#> [[15]]
#>  [1]  84 287  28   8   7  10  14  22  70  43  37  18 337  47  78  54  35  43  94
#> 
#> [[16]]
#>  [1]  78   0  96  45  36 100 107  35 277 248 163  30  17  22  29   0   2   6   0
#> 
#> [[17]]
#>  [1]  35   7  60  24  26  38  81  47 358 186 155  58  32  13  12   6   7   4   8
#> 
#> [[18]]
#>  [1]  57  90   4   5   1   5   5   3  31  44   3   4 378 328  13  66  37  18  39
#> 
#> [[19]]
#>  [1]  61  27  37  21  15  39  81  29 117 106  99  40 173  89  40  34  15  14  33
#> 
#> [[20]]
#>  [1]  37  53  45   4   9  39  32  18 139 105 133  19  29 115   7  26  35 108  22
#> 
#> [[21]]
#>  [1] 285  59  48   9   8  14  11  21  58  24  41  12 148  34  38  50  13  49  37
#> 
#> [[22]]
#>  [1]  32  14  76  28  56  13  12  68  70  87 245 128  15   1   2   5   2   6   3
#> 
#> [[23]]
#>  [1]  59  26  11  20  18  17  18  24  42  88  32  29  82  18 116  72  32  58  79
#> 
#> [[24]]
#>  [1] 146  56  40   7  12  22  11  21  22  24  59  22 106  19  32  52  12  53  45
#> 
#> [[25]]
#>  [1]  64  60   6   4   3   3   5  19   5  10   8   6 308  35  47  57  10  51  37
#> 
#> [[26]]
#>  [1]  67  94   2   3   1   3   5  35  33   8   6   5 185  10  43 101  45  35  50
#> 
#> [[27]]
#>  [1]  24   6  47  11   1   9  14  13  87 143  88  42  27  18  20  23  38  19  27
#> 
#> [[28]]
#>  [1]  39 138  31   4   4   9  10  30  12  20  34  15  98   8  44  20   6  48  82
#> 
#> [[29]]
#>  [1]  33 132   0   1   0   1   0   2   0   0   1   2 226   5  13  21  17  96  87
#> 
#> [[30]]
#>  [1]  26  44   6   4  11  14  11  33  23  12  39  25 135  16  40  45  11  40  60
#> 
#> [[31]]
#>  [1]  24  26  54  10  22   8  18  22  43  42 134  76  39   9   0  22   4   7  14
#> 
#> [[32]]
#>  [1]  49  48   3  14  12  13   9  38  49  19  21  12 104  12  24  21   9  31  50
#> 
#> [[33]]
#>  [1] 73 62 21  6 11 15 12 19 10  8 38 26 72  6 36 44  8 28 39
#> 
#> [[34]]
#>  [1]  35  45   3   4   1   1   0   5  20   5  11   5 136  11  41  46  13  50  95
#> 
#> [[35]]
#>  [1] 52 47  7 11  8  8 13 13 41 11 31 23 71 11 29 49 18 28 53
#> 
#> [[36]]
#>  [1]  62  49   9   5   4   3   8  13  13  14  52  16 150  10   8  26   8  41  23
#> 
#> [[37]]
#>  [1]  57  42   6   3   4   3   1   4  18  12   7   3 189   5  53  10   5  33  34
#> 
#> [[38]]
#>  [1] 27 27 41  4 32  2 12 27 19 49 54 34 36  4 18 22  7 23 31
#> 
#> [[39]]
#>  [1]  6 11 22  1  9 11  9 10 41 17 57 42 87 15 34 30  9 24 30
#> 
#> [[40]]
#>  [1] 26 28 22  9 14  4  5 29 37 45 80 26 37  5 10 29 11  5 14
#> 
#> [[41]]
#>  [1] 43 38 11  5  5  3  4 16 21 10 19 12 68  7 17 55  7 41 53
#> 
#> [[42]]
#>  [1]  14  59   2   4   4   4   4  15   5   5  13   5 112   5  32  42   4  29  62
#> 
#> [[43]]
#>  [1] 63 45  8  2  6  6  6  6 23 21 26 13 78  9 27 27  7 16 18
#> 
#> [[44]]
#>  [1]  2  3 54 10 34 32 40 89 37  9  3  0  2  3  2 14  2 24 37
#> 
#> [[45]]
#>  [1] 16 20 11  6 11  2  7 20 32 23 36 38 74  7 17 20  7 21 27
#> 
#> [[46]]
#>  [1] 10 21  7  1  2  4  5  6  8  9 19  6 71 15 33 36 28 37 52
#> 
#> [[47]]
#>  [1] 24 61  0  1  0  0  0  0  0  1  0  0 77 12 32 43 18 30 48
#> 
#> [[48]]
#>  [1]  8 20 19  2 12  3  5 12 45 29 57 19 63  3  6  7  7 10 16
#> 
#> [[49]]
#>  [1]  31  41   2   2   2   0   1   4   9   5   6   2 136   3   9  40  11  10  22
#> 
#> [[50]]
#>  [1] 42 58 15  2  3  2  0 15  8 11 35 23 77 24  6  5  1  8  1
#> 
#> [[51]]
#>  [1]  5 65  7  0  1  1  1 10  9 13 31 14 52  9 46 23  5 16 25
#> 
#> [[52]]
#>  [1] 47 24  4  7  8 14  8 16 28 20 32 13 36  2 17 21  4  5 16
#> 
#> [[53]]
#>  [1] 16 31 27  2  4  3  2 13 21 19 22 12 42  9 15 20  6 20 21
#> 
#> [[54]]
#>  [1]  3  9  5  7  7  9 18  4 29 13  8  7 51 37  6 21 26 21 22
#> 
#> [[55]]
#>  [1]  0  0 15  8  3 17 12 34 78 56 42 37  0  0  0  0  0  0  0
#> 
#> [[56]]
#>  [1]  8 18  2  2  0  1  1  3 16  1  1  2 38  3 18 24 18 47 97
#> 
#> [[57]]
#>  [1] 74 40  0  1  0  0  3  0 20  0  1  0 75 12  1  6 13  9 27
#> 
#> [[58]]
#>  [1]  0  0 30 22 21 14 15  5 49 80 32 13  0  0  0  0  0  0  0
#> 
#> [[59]]
#>  [1]  0  0  4  1  0  0  4 26 19 13 10 13  2  0 39 27  6 42 63
#> 
#> [[60]]
#>  [1] 12 34  4  2  3  4  8 10  3  7 12 10 72  7  8 12  2 27 26
#> 
#> [[61]]
#>  [1] 20 17 14  2  4  2  4 14  9 17 45 23 44  5  6  8  2 10  5
#> 
#> [[62]]
#>  [1]  4  9 15  1 13  5  4 12  5 14 60 17 29  6  7 12  6  7 11
#> 
#> [[63]]
#>  [1] 21  0 15 22  8 20  5  3 23 60 23 16  4 15  0  0  0  0  0
#> 
#> [[64]]
#>  [1] 18  3  5  6  3  4 12  2 26 23  9  6 54 18 10  4 10  8 13
#> 
#> [[65]]
#>  [1]  3  2 44  4 23  2  2 10 11 16 72 34  3  0  1  1  0  1  0
#> 
#> [[66]]
#>  [1]  6  4 23  5 16  3  9 11  8 25 50 36  9  1  8  5  1  2  6
#> 
#> [[67]]
#>  [1]  9 15  0  5  1  2  6 16 11 15 20 20 46  2  9 12  3 10 14
#> 
#> [[68]]
#>  [1]  5 12  1  1  1  1  3  7  6  2  8  4 19 21 12 26 11 31 44
#> 
#> [[69]]
#>  [1] 23  3 16  0  4  1  7  5 12 16 13  3 34 17 11  7 18  8 17
#> 
#> [[70]]
#>  [1] 37 33  5  0  0  0  1  4  7  6 17  0 87 10  0  2  1  2  2
#> 
#> [[71]]
#>  [1] 11  9  4  2  1  2  6  5 61 14 20  3 13  1  7 36  1  6 11
#> 
#> [[72]]
#>  [1]  1  3 26  5  7  3  4  8 24 18 62 46  1  0  0  1  0  0  1
#> 
#> [[73]]
#>  [1] 115   0   4   3   2   5  10   1  17  26  11   6   1   0   1   0   0   1   0
#> 
#> [[74]]
#>  [1] 22  9  4  0  2  4  7 15  8 12 21 10 19  0  3 17  7 14 29
#> 
#> [[75]]
#>  [1] 24 10  3  1  4  2  1 13  8  4  3  1 38  1 11 31  5 18 22
#> 
#> [[76]]
#>  [1] 12  2  3  1  1  5  6  0  8 18 17  4 26  5 21 17  9  1 33
#> 
#> [[77]]
#>  [1]  0 15  5  4  0  1  9 16 17  0 10  9 33  1  9 17  5 12 18
#> 
#> [[78]]
#>  [1]  5 14  9  2  7  0  2  7 32  5 16 12 10  3  7  9  4 11 10
#> 
#> [[79]]
#>  [1]  5  9  8  6  7  4  9  1 29 17 14  5 22  4  3  8  1  5  6
#> 
#> [[80]]
#>  [1]  0  1 11  1  5  6 11 13  4 18 36 16  9  0  5  6  3  3  6
#> 
#> [[81]]
#>  [1]  0 26  3  1  3  0  3  3 14 11  6  0  2 11 15 14 14 11  8
#> 
#> [[82]]
#>  [1]  3  3 15  2  6  5  3 12 28 12  7  8 21  2  2  5  2  5  3
#> 
#> [[83]]
#>  [1] 18  2 11  5  5  4  9  9  5 14 21  6  5  4  4  2  5  3 12
#> 
#> [[84]]
#>  [1]  2  0  7  3  6 11  8 11  5 19 39  6  8  0  3  7  0  2  4
#> 
#> [[85]]
#>  [1]  4 40  1  1  0  1  1  2  6  3  2  3 47  7  2  2  1  4  7
#> 
#> [[86]]
#>  [1]  0  3  9  2  8  0  4  7  6  7 24 11 11 11  5 10  5  3  7
#> 
#> [[87]]
#>  [1] 16  8  2  1  0  2  1 11 10 10 11  7 14  0  5 14  1 10 10
#> 
#> [[88]]
#>  [1]  0  4  9  1  3  3  5 12 10  9 22 11 14  4  5  5  1  7  6
#> 
#> [[89]]
#>  [1]  3  5  0  3  1  1  2 16 20  6  9 11 10  0  6 10  0  8  5
#> 
#> [[90]]
#>  [1]  3  4  2  1  1  1  0  1  5  1  7  2 35 19 15  9  3  2  2
#> 
#> [[91]]
#>  [1] 16  2  1  2  0  0  2  8  2  2  5  4 15  0  2 17  3 10 17
#> 
#> [[92]]
#>  [1] 23 14  0  0  0  0  0  1  0  1  0  0 25  9  1 12  5  8  9
#> 
#> [[93]]
#>  [1]  6  6  3  2  0  1  3  0  6  6  7  1 34  8  1  1  2  7 10
#> 
#> [[94]]
#>  [1]  2  8  1  0  0  0  0  1  0  0  1  0 24  5  8 29  7 11  7
#> 
#> [[95]]
#>  [1] 10  5  3  0  1  2  2  2 11  6  8  8 12  2  6  8  0  8  7
#> 
#> [[96]]
#>  [1] 32 10  0  0  0  0  2  0  0  3  2  3 34  2  2  2  0  2  2
#> 
#> [[97]]
#>  [1]  5 20  6  0  1  1  1  3  0  2 16  5 16  2  2  3  2  7  2
#> 
#> [[98]]
#>  [1]  1  0  1  0  1  2  2  4  9 23 34  7  0  0  1  1  0  2  4
#> 
#> [[99]]
#>  [1]  0  0  9  0  2  0  1 10  7  9 29 17  0  0  0  1  0  2  2
#> 
#> [[100]]
#>  [1]  0  0  3  6  2  8 14 10  9  8  4  6  0  0  0  0  2 10  5
#> 
#> [[101]]
#>  [1] 14  5  3  0  1  2  1 10  6  5  3  4  9  1  7  2  1  3  7
#> 
#> [[102]]
#>  [1]  1  5  3  0  3  1  1  2  3  8 13  4 17  3  4  3  1  4  8
#> 
#> [[103]]
#>  [1]  5 11  0  1  2  1  2  2  4  4  5  5  5  0  3 14  0  3 16
#> 
#> [[104]]
#>  [1] 65  3  0  0  0  0  0  0  0  0  0  0 13  1  0  0  0  0  0
#> 
#> [[105]]
#>  [1]  9  0  3  6  4  0  6 16 11 11  8  4  3  0  0  0  0  0  0
#> 
#> [[106]]
#>  [1]  1  0  5  1  3  3  3  3  0  9 14  2  2  0  2  3  4  8 14
#> 
#> [[107]]
#>  [1]  1  0  0  0  1  0  3  0  2  1  2  2  5  4  2 15 12 14 12
#> 
#> [[108]]
#>  [1]  3  6  3  1  0  1  0  3  1  3  8  2  7  4  8  2  2 13  9
#> 
#> [[109]]
#>  [1]  3  0  5  0  5  0  2  2  1 14 22  5  6  1  1  1  0  3  2
#> 
#> [[110]]
#>  [1]  0  0  0  1  0  1  1  7 11  7 27 18  0  0  0  0  0  0  0
#> 
#> [[111]]
#>  [1]  3  7  0  0  0  0  2  1  2  1  2  2 19  0  5  7  3 11  6
#> 
#> [[112]]
#>  [1]  5  1  3  0  3  1  4  2  5  6  4  2  5  0  4  4  1 10  8
#> 
#> [[113]]
#>  [1]  9  6  0  0  0  1  1 11  3  0  3  1  7  3  4  6  0  5  4
#> 
#> [[114]]
#>  [1]  2  3  1  0  1  1  3  4  7  7 11  5  7  2  2  2  2  2  0
#> 
#> [[115]]
#>  [1] 3 9 0 1 1 0 2 3 4 3 6 2 9 1 3 3 1 4 5
#> 
#> [[116]]
#>  [1]  0  0  0  3  9  3 15  1 11 11  3  2  0  0  0  0  0  0  0
#> 
#> [[117]]
#>  [1]  5  0  4  1  2  2  4  4  2  7 16  4  3  0  0  1  1  0  1
#> 
#> [[118]]
#>  [1]  8  6  0  0  1  1  2  0  0  4  1  1 11  1  2  5  2  5  6
#> 
#> [[119]]
#>  [1] 7 3 2 0 0 6 6 5 1 4 9 7 1 1 0 1 0 0 1
#> 
#> [[120]]
#>  [1] 13  7  0  2  0  5  1  4  0  2  2  1  3  0  1  5  0  5  3
#> 
#> [[121]]
#>  [1] 0 5 2 0 0 2 1 4 1 4 2 1 8 1 7 5 2 3 4
#> 
#> [[122]]
#>  [1]  1  0  3  4  2  0  4  2  6 13  9  1  0  0  0  4  0  0  1
#> 
#> [[123]]
#>  [1] 13  5  1  0  0  2  0  0  3  3  3  2  6  0  1  1  2  2  5
#> 
#> [[124]]
#>  [1] 4 2 1 1 0 2 1 1 6 5 3 2 7 0 1 2 1 3 7
#> 
#> [[125]]
#>  [1] 3 0 0 0 0 0 4 5 6 0 1 0 5 2 2 7 1 2 9
#> 
#> [[126]]
#>  [1] 0 0 3 0 1 1 3 4 5 3 7 2 1 0 3 2 0 5 7
#> 
#> [[127]]
#>  [1]  4  3  0  0  0  0  0  2  0  2  0  0 20  0  1  3  2  1  7
#> 
#> [[128]]
#>  [1] 0 2 2 1 0 1 3 2 4 7 8 6 4 0 1 1 0 2 1
#> 
#> [[129]]
#>  [1] 5 0 1 3 1 3 4 0 2 3 6 5 5 0 0 2 0 3 1
#> 
#> [[130]]
#>  [1]  1  3  0  0  0  0  0  0  0  1  0  0  6  0  8  7  0  2 15
#> 
#> [[131]]
#>  [1] 0 3 2 0 1 0 1 2 6 2 3 2 6 2 1 2 2 4 4
#> 
#> [[132]]
#>  [1] 3 3 0 0 0 1 0 2 6 0 7 6 5 1 2 2 1 0 4
#> 
#> [[133]]
#>  [1]  0  1  0  0  0  0  1  0  4  1  3  1 12  2  3  3  0  5  6
#> 
#> [[134]]
#>  [1]  0  0  2  2  3  0  2  4 10  5  5  2  1  0  0  0  0  3  2
#> 
#> [[135]]
#>  [1] 0 0 1 1 0 1 3 0 8 2 6 2 0 9 3 2 0 0 1
#> 
#> [[136]]
#>  [1]  1  3  1  1  0  0  0  1  1  1  1  1  5  0  1  2  2  3 14
#> 
#> [[137]]
#>  [1] 5 0 2 2 0 1 0 1 3 7 3 1 2 5 1 1 1 2 1
#> 
#> [[138]]
#>  [1] 2 0 1 1 0 3 5 2 7 5 1 0 4 2 1 0 2 0 1
#> 
#> [[139]]
#>  [1]  0  4  0  1  0  1  0  2  2  0  1  0 10  2  0  4  1  4  4
#> 
#> [[140]]
#>  [1]  0  0  1  0  1  0  2  0 12 10  3  1  0  0  2  0  4  0  0
#> 
#> [[141]]
#>  [1] 1 0 0 1 0 1 2 3 4 3 0 3 5 1 1 7 2 0 1
#> 
#> [[142]]
#>  [1] 0 0 2 0 2 6 1 9 0 5 1 9 0 0 0 0 0 0 0
#> 
#> [[143]]
#>  [1]  7  0  0  0  0  0  0  1  1  1  1  0  2  0  0  6  3  3 10
#> 
#> [[144]]
#>  [1] 19  0  1  0  0  0  0  0  0  8  4  2  1  0  0  0  0  0  0
#> 
#> [[145]]
#>  [1] 3 2 0 0 0 0 1 1 4 0 1 1 6 1 4 3 2 3 1
#> 
#> [[146]]
#>  [1] 5 0 3 0 2 1 1 2 2 4 6 1 1 0 0 1 0 1 1
#> 
#> [[147]]
#>  [1] 1 1 0 1 0 0 1 4 2 3 1 1 4 1 2 1 0 5 3
#> 
#> [[148]]
#>  [1] 4 8 0 1 0 0 0 0 0 0 0 0 3 0 1 2 0 6 5
#> 
#> [[149]]
#>  [1] 0 0 6 0 1 1 3 0 2 2 9 0 1 0 0 1 0 0 2
#> 
#> [[150]]
#>  [1] 5 0 0 1 1 2 6 3 2 4 0 1 0 0 0 0 0 1 1
#> 
#> [[151]]
#>  [1] 5 6 0 0 0 0 1 0 0 1 0 0 3 4 0 2 1 2 1
#> 
#> [[152]]
#>  [1] 0 0 1 1 3 3 4 2 4 3 3 1 0 0 0 0 0 0 0
#> 
#> [[153]]
#>  [1] 0 5 1 0 1 0 0 5 2 1 1 1 4 1 2 1 0 0 0
#> 
#> [[154]]
#>  [1] 3 3 1 0 0 1 2 2 2 3 2 0 0 0 1 0 0 2 2
#> 
#> [[155]]
#>  [1] 0 3 0 1 0 0 0 3 1 0 0 0 3 1 1 3 1 3 4
#> 
#> [[156]]
#>  [1] 2 2 0 0 0 0 0 1 0 0 1 0 2 1 1 4 0 3 6
#> 
#> [[157]]
#>  [1] 1 3 3 0 1 2 1 0 2 0 1 1 4 0 2 0 0 1 1
#> 
#> [[158]]
#>  [1] 0 1 1 0 1 1 0 2 3 1 2 1 2 1 0 2 1 1 3
#> 
#> [[159]]
#>  [1] 0 4 2 0 0 0 1 1 2 1 0 1 3 2 1 1 0 1 2
#> 
#> [[160]]
#>  [1] 3 0 0 0 0 0 0 1 0 0 5 1 3 0 1 7 0 0 1
#> 
#> [[161]]
#>  [1] 2 1 0 1 1 0 0 0 2 1 0 1 3 1 1 4 0 1 2
#> 
#> [[162]]
#>  [1] 2 2 1 0 1 0 0 1 0 1 2 0 2 1 1 3 0 1 2
#> 
#> [[163]]
#>  [1] 2 3 0 0 0 0 0 2 4 0 1 0 0 0 1 0 1 2 3
#> 
#> [[164]]
#>  [1] 11  1  0  0  0  1  0  0  1  1  0  1  2  0  1  0  0  0  0
#> 
#> [[165]]
#>  [1] 0 0 4 1 0 2 1 1 2 1 4 2 0 0 0 0 0 0 0
#> 
#> [[166]]
#>  [1] 8 0 0 0 0 0 0 0 0 0 2 1 0 0 0 3 0 1 3
#> 
#> [[167]]
#>  [1] 0 0 0 0 0 0 2 0 0 0 0 0 0 3 4 6 1 0 2
#> 
#> [[168]]
#>  [1] 10  0  0  0  0  0  0  0  0  0  0  0  7  0  0  0  0  0  0
#> 
#> [[169]]
#>  [1] 1 0 0 0 1 2 1 0 3 6 0 1 0 0 0 1 0 0 1
#> 
#> [[170]]
#>  [1] 1 4 1 0 0 1 2 0 0 2 0 1 2 0 1 1 0 0 1
#> 
#> [[171]]
#>  [1] 0 0 1 0 0 0 3 0 1 3 5 1 0 0 0 0 0 0 2
#> 
#> [[172]]
#>  [1] 3 0 0 1 0 0 1 1 2 0 0 1 2 1 2 0 0 0 2
#> 
#> [[173]]
#>  [1] 0 0 0 4 1 1 0 0 3 5 2 0 0 0 0 0 0 0 0
#> 
#> [[174]]
#>  [1] 1 0 1 2 0 0 0 1 4 2 0 0 0 1 0 0 1 2 1
#> 
#> [[175]]
#>  [1] 1 0 2 0 0 0 2 0 1 0 1 1 3 0 4 1 0 0 0
#> 
#> [[176]]
#>  [1] 2 1 1 1 0 0 1 1 1 1 3 0 2 1 0 0 0 0 0
#> 
#> [[177]]
#>  [1] 0 0 0 0 0 0 0 0 7 4 3 1 0 0 0 0 0 0 0
#> 
#> [[178]]
#>  [1] 4 0 2 0 2 1 0 4 0 0 0 0 0 0 0 0 0 2 0
#> 
#> [[179]]
#>  [1] 0 0 0 0 2 1 0 0 7 2 2 0 0 0 0 0 0 0 0
#> 
#> [[180]]
#>  [1] 0 0 0 0 0 2 0 0 3 1 1 1 3 1 1 0 0 0 1
#> 
#> [[181]]
#>  [1] 0 0 1 0 0 1 0 0 1 0 9 2 0 0 0 0 0 0 0
#> 
#> [[182]]
#>  [1] 5 0 0 0 0 0 0 0 1 0 0 0 0 0 1 0 4 1 2
#> 
#> [[183]]
#>  [1] 2 0 1 2 0 0 0 1 1 0 1 1 1 0 0 1 0 1 1
#> 
#> [[184]]
#>  [1] 0 0 0 0 1 1 0 0 2 0 0 0 6 1 0 1 0 1 0
#> 
#> [[185]]
#>  [1] 0 0 2 1 1 0 1 1 0 1 5 1 0 0 0 0 0 0 0
#> 
#> [[186]]
#>  [1] 2 0 0 0 1 0 0 0 2 1 2 2 0 2 0 0 0 0 1
#> 
#> [[187]]
#>  [1] 1 0 0 1 1 1 1 0 3 1 0 0 0 0 1 0 1 1 0
#> 
#> [[188]]
#>  [1] 1 0 0 0 0 0 0 1 2 1 4 1 0 0 0 0 0 2 0
#> 
#> [[189]]
#>  [1] 0 0 2 1 1 0 0 1 0 0 1 0 0 1 0 1 1 0 3
#> 
#> [[190]]
#>  [1] 1 0 1 3 0 0 0 0 3 0 0 0 2 0 0 0 0 1 1
#> 
#> [[191]]
#>  [1] 0 0 0 0 0 1 0 0 2 2 1 4 0 0 0 0 1 1 0
#> 
#> [[192]]
#>  [1] 0 0 0 0 2 1 3 0 2 3 1 0 0 0 0 0 0 0 0
#> 
#> [[193]]
#>  [1] 0 0 0 0 0 1 0 0 1 0 2 0 0 0 0 2 0 3 2
#> 
#> [[194]]
#>  [1] 0 0 0 0 1 1 1 0 2 1 0 0 1 0 0 2 2 0 0
#> 
#> [[195]]
#>  [1] 0 0 0 0 0 2 2 2 5 0 0 0 0 0 0 0 0 0 0
#> 
#> [[196]]
#>  [1] 1 1 0 0 0 0 0 0 2 0 3 2 2 0 0 0 0 0 0
#> 
#> [[197]]
#>  [1] 1 0 1 0 0 1 0 0 2 1 1 3 0 0 0 1 0 0 0
#> 
#> [[198]]
#>  [1] 0 0 2 0 0 8 0 0 0 0 0 1 0 0 0 0 0 0 0
#> 
#> [[199]]
#>  [1] 0 0 1 0 0 0 0 0 1 0 0 0 4 0 0 0 1 1 2
#> 
#> [[200]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 3 7
#> 
#> [[201]]
#>  [1] 0 2 0 0 0 0 0 0 0 0 0 0 3 2 1 0 1 1 0
#> 
#> [[202]]
#>  [1] 3 1 0 0 0 0 1 0 1 0 0 1 2 0 0 0 0 0 0
#> 
#> [[203]]
#>  [1] 1 1 0 0 0 1 0 1 0 0 1 0 2 1 0 0 0 1 0
#> 
#> [[204]]
#>  [1] 0 4 1 0 0 0 0 3 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[205]]
#>  [1] 0 1 0 0 0 0 0 1 1 1 0 0 1 1 2 0 1 0 0
#> 
#> [[206]]
#>  [1] 3 0 0 0 0 0 0 0 0 0 0 0 3 0 0 1 0 1 1
#> 
#> [[207]]
#>  [1] 4 1 0 0 0 0 0 0 1 1 0 0 2 0 0 0 0 0 0
#> 
#> [[208]]
#>  [1] 3 0 1 0 0 0 1 0 0 0 1 0 2 0 0 1 0 0 0
#> 
#> [[209]]
#>  [1] 0 0 1 1 0 0 1 0 2 1 2 0 1 0 0 0 0 0 0
#> 
#> [[210]]
#>  [1] 0 0 1 0 2 1 5 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[211]]
#>  [1] 4 1 0 0 0 0 0 0 0 0 0 0 2 0 0 0 0 0 2
#> 
#> [[212]]
#>  [1] 0 0 2 0 0 1 0 0 4 1 0 0 0 0 0 0 0 0 0
#> 
#> [[213]]
#>  [1] 0 0 0 1 0 0 1 0 1 2 0 0 0 2 0 1 0 0 0
#> 
#> [[214]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 1 0 3 0 0 3 0 0
#> 
#> [[215]]
#>  [1] 1 2 0 0 1 0 0 0 0 0 0 0 3 0 0 1 0 0 0
#> 
#> [[216]]
#>  [1] 2 0 0 0 0 0 0 0 0 0 0 0 3 1 0 0 0 0 2
#> 
#> [[217]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 5 0 2 0 0 1 0
#> 
#> [[218]]
#>  [1] 0 0 0 2 0 0 0 0 1 4 1 0 0 0 0 0 0 0 0
#> 
#> [[219]]
#>  [1] 0 0 1 0 0 0 0 0 5 1 0 0 0 0 0 0 0 0 0
#> 
#> [[220]]
#>  [1] 3 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 0 2 0
#> 
#> [[221]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 4 0 0 0 1 1 0
#> 
#> [[222]]
#>  [1] 5 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0
#> 
#> [[223]]
#>  [1] 0 0 1 0 0 1 2 1 0 0 1 0 1 0 0 0 0 0 0
#> 
#> [[224]]
#>  [1] 1 0 0 0 1 0 0 0 0 0 2 0 1 0 0 1 0 0 1
#> 
#> [[225]]
#>  [1] 1 1 0 0 0 0 0 0 2 0 3 0 0 0 0 0 0 0 0
#> 
#> [[226]]
#>  [1] 0 0 0 0 0 1 1 1 1 2 0 1 0 0 0 0 0 0 0
#> 
#> [[227]]
#>  [1] 2 0 0 2 0 0 0 1 0 0 0 0 0 0 1 0 0 1 0
#> 
#> [[228]]
#>  [1] 0 0 0 0 0 0 0 0 2 0 0 1 0 0 0 0 1 0 2
#> 
#> [[229]]
#>  [1] 0 0 0 1 0 2 3 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[230]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 2 0 1 1 0 2 0
#> 
#> [[231]]
#>  [1] 0 0 1 0 0 1 0 0 0 2 1 0 1 0 0 0 0 0 0
#> 
#> [[232]]
#>  [1] 0 0 0 0 0 1 0 0 3 1 1 0 0 0 0 0 0 0 0
#> 
#> [[233]]
#>  [1] 0 0 0 1 0 0 0 0 1 0 0 1 0 0 0 1 2 0 0
#> 
#> [[234]]
#>  [1] 0 0 0 0 1 1 1 2 0 0 0 0 0 0 1 0 0 0 0
#> 
#> [[235]]
#>  [1] 0 0 0 0 0 0 2 0 2 1 1 0 0 0 0 0 0 0 0
#> 
#> [[236]]
#>  [1] 0 0 0 0 0 0 0 0 5 1 0 0 0 0 0 0 0 0 0
#> 
#> [[237]]
#>  [1] 2 0 0 0 0 1 0 0 0 0 1 0 2 0 0 0 0 0 0
#> 
#> [[238]]
#>  [1] 0 1 0 0 0 0 0 0 2 0 0 0 1 0 0 0 0 2 0
#> 
#> [[239]]
#>  [1] 0 0 0 0 0 0 0 0 1 1 2 1 0 0 0 1 0 0 0
#> 
#> [[240]]
#>  [1] 0 0 0 0 1 2 0 0 1 0 1 0 0 1 0 0 0 0 0
#> 
#> [[241]]
#>  [1] 0 0 0 1 0 1 0 0 4 0 0 0 0 0 0 0 0 0 0
#> 
#> [[242]]
#>  [1] 0 0 1 1 0 0 0 1 0 0 1 1 0 0 0 0 0 0 0
#> 
#> [[243]]
#>  [1] 1 0 0 0 0 0 0 1 0 0 0 0 0 0 0 1 0 1 1
#> 
#> [[244]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 1 3 0 0 0 0 0 0
#> 
#> [[245]]
#>  [1] 0 0 0 1 0 1 0 0 0 1 2 0 0 0 0 0 0 0 0
#> 
#> [[246]]
#>  [1] 0 0 0 1 0 0 1 0 0 0 2 0 0 0 0 0 0 1 0
#> 
#> [[247]]
#>  [1] 0 0 0 0 0 0 0 1 1 1 0 0 1 0 1 0 0 0 0
#> 
#> [[248]]
#>  [1] 0 0 0 0 1 0 1 0 1 0 0 0 0 0 0 0 0 1 1
#> 
#> [[249]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 3 1 0 0 0 0 0
#> 
#> [[250]]
#>  [1] 0 0 1 1 0 1 1 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[251]]
#>  [1] 0 3 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0
#> 
#> [[252]]
#>  [1] 0 1 0 0 0 1 0 1 0 0 0 0 0 0 0 1 0 0 1
#> 
#> [[253]]
#>  [1] 0 0 0 0 0 0 1 0 0 1 0 1 0 0 0 1 0 0 0
#> 
#> [[254]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 1 1 1 0 0 0 0 0 0
#> 
#> [[255]]
#>  [1] 0 0 1 1 0 0 2 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[256]]
#>  [1] 2 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 1 0
#> 
#> [[257]]
#>  [1] 0 1 0 0 0 0 0 1 1 0 0 0 1 0 0 0 0 0 0
#> 
#> [[258]]
#>  [1] 0 0 0 0 0 1 3 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[259]]
#>  [1] 1 0 0 1 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0
#> 
#> [[260]]
#>  [1] 0 0 0 0 0 0 0 1 0 2 0 0 1 0 0 0 0 0 0
#> 
#> [[261]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 2 1 0 0 0 0 0 0 0
#> 
#> [[262]]
#>  [1] 0 0 0 0 1 0 1 0 0 0 0 0 2 0 0 0 0 0 0
#> 
#> [[263]]
#>  [1] 0 0 0 0 0 0 0 0 4 0 0 0 0 0 0 0 0 0 0
#> 
#> [[264]]
#>  [1] 0 0 0 0 2 0 0 0 0 0 0 0 0 0 2 0 0 0 0
#> 
#> [[265]]
#>  [1] 0 0 0 0 0 2 0 1 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[266]]
#>  [1] 0 0 0 0 0 0 0 2 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[267]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 1 0 0 0 1 0
#> 
#> [[268]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 2 0 0 0 0 1 0
#> 
#> [[269]]
#>  [1] 0 0 0 0 0 0 1 0 1 1 0 0 0 0 0 0 0 0 0
#> 
#> [[270]]
#>  [1] 0 0 0 0 0 1 0 0 1 1 0 0 0 0 0 0 0 0 0
#> 
#> [[271]]
#>  [1] 0 1 0 0 0 0 0 0 2 0 0 0 0 0 0 0 0 0 0
#> 
#> [[272]]
#>  [1] 0 1 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0
#> 
#> [[273]]
#>  [1] 0 0 0 0 1 0 0 0 0 1 1 0 0 0 0 0 0 0 0
#> 
#> [[274]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 2 0 0 0 0 0 0
#> 
#> [[275]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 2
#> 
#> [[276]]
#>  [1] 0 0 0 0 0 0 0 0 2 0 1 0 0 0 0 0 0 0 0
#> 
#> [[277]]
#>  [1] 1 0 1 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[278]]
#>  [1] 0 0 0 0 0 0 1 0 0 1 0 1 0 0 0 0 0 0 0
#> 
#> [[279]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 2 0 0 0 0 0 0 0
#> 
#> [[280]]
#>  [1] 0 0 0 0 0 0 1 0 2 0 0 0 0 0 0 0 0 0 0
#> 
#> [[281]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 3 0 0 0 0 0 0
#> 
#> [[282]]
#>  [1] 0 0 1 0 0 0 0 0 1 0 0 0 0 1 0 0 0 0 0
#> 
#> [[283]]
#>  [1] 0 0 0 0 0 0 0 1 2 0 0 0 0 0 0 0 0 0 0
#> 
#> [[284]]
#>  [1] 0 0 0 0 0 0 0 0 2 0 0 1 0 0 0 0 0 0 0
#> 
#> [[285]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 1 0 1 0 0 0 0 0
#> 
#> [[286]]
#>  [1] 0 0 0 0 0 2 0 0 0 0 0 0 0 0 0 0 1 0 0
#> 
#> [[287]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 2 0 0 0 0 0 0 0 0
#> 
#> [[288]]
#>  [1] 0 0 0 1 0 1 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[289]]
#>  [1] 0 1 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[290]]
#>  [1] 1 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 1 0
#> 
#> [[291]]
#>  [1] 0 0 0 0 0 0 2 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[292]]
#>  [1] 0 0 0 0 0 0 0 0 1 1 0 0 1 0 0 0 0 0 0
#> 
#> [[293]]
#>  [1] 0 0 0 0 1 0 0 0 0 1 1 0 0 0 0 0 0 0 0
#> 
#> [[294]]
#>  [1] 0 2 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[295]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0
#> 
#> [[296]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 1 0
#> 
#> [[297]]
#>  [1] 0 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[298]]
#>  [1] 0 0 1 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[299]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[300]]
#>  [1] 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[301]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[302]]
#>  [1] 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[303]]
#>  [1] 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[304]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 1 0 0 0 0
#> 
#> [[305]]
#>  [1] 0 0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[306]]
#>  [1] 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[307]]
#>  [1] 0 0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[308]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[309]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 0
#> 
#> [[310]]
#>  [1] 1 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[311]]
#>  [1] 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[312]]
#>  [1] 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[313]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 1
#> 
#> [[314]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 2
#> 
#> [[315]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[316]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0
#> 
#> [[317]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[318]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[319]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[320]]
#>  [1] 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 0 0
#> 
#> [[321]]
#>  [1] 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[322]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[323]]
#>  [1] 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[324]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 0 0
#> 
#> [[325]]
#>  [1] 0 0 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[326]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0
#> 
#> [[327]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[328]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[329]]
#>  [1] 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[330]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[331]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[332]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[333]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[334]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[335]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[336]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[337]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[338]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[339]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[340]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[341]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[342]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[343]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0
#> 
#> [[344]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[345]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[346]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[347]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[348]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0
#> 
#> [[349]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[350]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[351]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[352]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[353]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[354]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[355]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[356]]
#>  [1] 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[357]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[358]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[359]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[360]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[361]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[362]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[363]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[364]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[365]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[366]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> 
#> [[367]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0
#> 
#> [[368]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[369]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0
#> 
#> [[370]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[371]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[372]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[373]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[374]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[375]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[376]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[377]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0
#> 
#> [[378]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0
#> 
#> [[379]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0
#> 
#> [[380]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[381]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[382]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[383]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[384]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[385]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[386]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[387]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[388]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[389]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[390]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[391]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0
#> 
#> [[392]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[393]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[394]]
#>  [1] 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[395]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0
#> 
#> [[396]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[397]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0
#> 
#> [[398]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0
#> 
#> [[399]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[400]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[401]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[402]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[403]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[404]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[405]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[406]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[407]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[408]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[409]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[410]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[411]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[412]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[413]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[414]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[415]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[416]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[417]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> 
#> [[418]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[419]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[420]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[421]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[422]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[423]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[424]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[425]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[426]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[427]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[428]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[429]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[430]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[431]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[432]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> 
#> [[433]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[434]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[435]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[436]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[437]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[438]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[439]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[440]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[441]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[442]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[443]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[444]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[445]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[446]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[447]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> 
#> [[448]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[449]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[450]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[451]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[452]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[453]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[454]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[455]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[456]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[457]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[458]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[459]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[460]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[461]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[462]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0
#> 
#> [[463]]
#>  [1] 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[464]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[465]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0
#> 
#> [[466]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[467]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[468]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[469]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[470]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[471]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[472]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[473]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[474]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[475]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[476]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[477]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[478]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[479]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[480]]
#>  [1] 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[481]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[482]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[483]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[484]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[485]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[486]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[487]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[488]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0
#> 
#> [[489]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[490]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[491]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[492]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0
#> 
#> [[493]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[494]]
#>  [1] 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[495]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[496]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[497]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[498]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[499]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1
#> 
#> [[500]]
#>  [1] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0
#> 
#> [[501]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[502]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[503]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[504]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[505]]
#>  [1] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[506]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[507]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> 
#> [[508]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[509]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[510]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[511]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[512]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[513]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[514]]
#>  [1] 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[515]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[516]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[517]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[518]]
#>  [1] 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[519]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[520]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[521]]
#>  [1] 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[522]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> 
#> [[523]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> 
#> [[524]]
#>  [1] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[525]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0
#> 
#> [[526]]
#>  [1] 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#> 
#> [[527]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0
#> 
#> [[528]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[529]]
#>  [1] 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[530]]
#>  [1] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> [[531]]
#>  [1] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>