new_dataset

Create a pointer to an instance of the 'Dataset' c++ class.

The 'Dataset' class is the c++ implementation of the R6 'dataset' object. This class allows package developers access to additional functionality. 'Dataset' stores nucleotide sequences, abundance, sample and treatment assignments, taxonomic classifications, asv / otu clusters. It creates various reports and summaries. It is designed to facilitate data transfer and access across multiple R packages.

new_dataset(dataset_name, processors)

Arguments

dataset_name

string, a string containing the dataset name

processors

integer, number of cores to use during summary functions.

Value

Rcpp::XPtr<Dataset> pointer to an instance of the 'Dataset' c++ class.

See also

Examples

# to create a new dataset and allow for 4 processors during summary,
# run the following:

dataset <- new_dataset("soil", 4)