read_mothur_cons_taxonomy

Read a mothur formatted cons_taxonomy file

read_mothur_cons_taxonomy(taxonomy)

Arguments

taxonomy

file name, a mothur consensus taxonomy file. The cons_taxonomy file is created by classify.otu.

Value

A data.frame containing the bin names, bin abundances and bin taxonomies.

Examples

# You can add the otu assignments and bin taxonomies to the your data set
# using the following:

otu_data <- read_mothur_cons_taxonomy(rdataset_example(
  "final.cons.taxonomy"
))

data <- new_dataset()
assign_bins(data, otu_data)
#> ℹ Assigned 531 otu bins.
#> [1] 531
assign_bin_taxonomy(data, otu_data)
#> ℹ Assigned 531 otu bin taxonomies.
#> [1] 531