read_qiime2

The read_qiime function reads various types of .qza files created by qiime2, and creates a 'dataset' object.

To generate the various input files you can follow qiime moving-pictures.

read_qiime2(
  fasta = NULL,
  taxonomy = NULL,
  biom = NULL,
  cons_fasta = NULL,
  cons_taxonomy = NULL,
  tree = NULL,
  metadata = NULL,
  dataset_name = "",
  dir_path = NULL,
  remove_unpacked_artifacts = TRUE
)

Arguments

fasta

filename, a .qza file containing DNA sequence strings.

taxonomy

filename, a .qza file containing classifications.

biom

filename, a .qza file containing a feature table in biom format. This will be used to assign bin abundance and sample data.

cons_fasta

filename, a .qza file containing DNA sequence strings of representative sequences for each bin.

cons_taxonomy

filename, a .qza file containing conscensus classifications for each bin.

tree

filename, a .qza file containing a tree.

metadata

filename, a .tsv file containing metadata

dataset_name

A string containing a name for your dataset.

dir_path

a string containing the name of directory where the artifacts files should be unpacked. Default = current working directory.

remove_unpacked_artifacts

boolean, When TRUE, the unpacked artifacts and temporary directories will be removed. Default = TRUE.

Value

A 'dataset' object

Examples

# Using the example files from moving-pictures

# dataset <- read_qiime2(
#   cons_fasta = rdataset_example("rep_seqs.qza"),
#   biom = rdataset_example("table.qza"),
#   cons_taxonomy = rdataset_example("taxonomy.qza"),
#   tree = rdataset_example("rooted-tree.qza"),
#   metadata = rdataset_example("sample_metadata.tsv"),
#   dataset_name = "qiime_moving_pictures"
# )